+++ Runtime/IDataConfigTable.cs
@@ -2,6 +2,8 @@ namespace Dreamy.DataConfig
 {
     public interface IDataConfigTable
     {
+        string DocumentName { get; }
+
         int SchemaVersion { get; }
 
         void Initialize(string documentName);